[core] Move type declarations to `chains/interfaces`
